Refactor core metrics into gwvet.metric.core
